requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
certifi==2024.8.30
charset-normalizer==3.3.2
decorator==4.4.2
Django==5.1.1
djangorestframework==3.15.2
frozenlist==1.4.1
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
moviepy==1.0.3
multidict==6.1.0
numpy==2.1.1
opencage==3.0.2
pillow==10.4.0
proglog==0.1.10
python-dotenv==1.0.1
python-magic==0.4.27
requests==2.32.3
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.13.1
